future-ready-hackathon/backend/requirements.txt
fastapi
uvicorn
pydantic
llama-parse==0.6.60
llama-index==0.13.2
python-multipart

# LangChain dependencies for LLM-based scoring
langchain-core>=0.1.0
langchain-openai>=0.1.0
langchain-google-genai>=1.0.0
aiofiles

# Environment variable support
python-dotenv